Converted BigDecimal to double when transferring from product dto

+++ project1-backend/src/main/java/com/revature/controllers/ProductController.java
@@ -58,7 +58,7 @@ public class ProductController {
         Product newProd=new Product();
         newProd.setName(        productDTO.getName());
         newProd.setDescription( productDTO.getDescription());
-        newProd.setPrice(       productDTO.getPrice());
+        newProd.setPrice(       productDTO.getPrice().doubleValue());
         newProd.setCategory(existingCategory);
 
         Product saved = ps.addProduct(newProd);
